added images folder and edited recipe related files

+++ controllers/recipeController.js
@@ -12,9 +12,20 @@ exports.getRecipes = async (req,res)=>{
     }
     catch(error){
         console.log(error);
-        res.render("error",{error:"Error in loading recipes"})
+        return res.render("error",{error:"Error in loading recipes"})
     }
 }
+
+
+exports.showAddform = (req,res)=>{
+    try{
+        return res.render("add-recipe")
+    }
+    catch(error){
+        return res.send("Error in add form")
+    }
+}
+
 //create
 exports.createRecipes = async (req,res) =>{
     try{
@@ -22,23 +33,47 @@ exports.createRecipes = async (req,res) =>{
         const ingredients = req.body.ingredients
         const instructions = req.body.instructions
         //
-        const createdBy = req.user.id
+        const createdBy = req.user.email
+        const image = req.body.image || "/images/default.jpg";
+
         const data = {
             title: title,
             ingredients: ingredients,
             instructions: instructions,
-            image: "/images/default.jpg",
+            image: image,
             createdBy: createdBy
         }
+
         //okay this part creates a new recipe to be added 
         //new recipes will have just the default image 
         await Recipe.createRecipe(data)
-        //this will redirect back to the main page, so we can see our change to recipes
-        res.redirect("/recipe")
+
+        // 
+        return res.redirect("/recipe")
     }
     catch(error){
         console.log(error)
-        res.render("error", { error: "Error creating recipe" })
+        return res.render("error", { error: "Error creating recipe" })
+    }
+}
+
+// Edit form, we will populate it with the preexisting data
+//make sure that the recipe can be found
+//then render the edit-recipe page
+exports.showEditForm = async (req,res)=>{
+    try{
+        // fetch recipe so form can be pre-filled
+        const recipe = await Recipe.findById(req.params.id)
+
+        if (!recipe){
+            return res.send("Recipe not found")
+        }
+
+        //  pass recipe to EJS
+        return res.render("edit-recipe", { recipe })
+    }
+    catch(error){
+        return res.send("There has been an error rendering edit recipe ")
     }
 }
 
@@ -47,34 +82,49 @@ exports.updateRecipes = async (req,res) => {
     try{
         //I'll be using params instead of query, because i plan to use /:id not ?id= etc 
         const id = req.params.id;
+
+        // check if its a valid MongoDB ID (prevents crash)
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.send("Invalid ID")
+        }
+
         //maybe i should make a new recipe 
         //check the recipe id first
         const recipe = await Recipe.findById(id)
+
         if (!recipe){
-            res.send("No ID has been found")
+            return res.send("No ID has been found")
         }
+
         //check ownership(havent learn authentication so ill move on )
         // placeholder()
+        if (recipe.createdBy !== req.user.email){
+            return res.send("Not allowed")
+        }
+
         //take data from add recipe 
         const title = req.body.title
         const ingredients = req.body.ingredients
         const instructions = req.body.instructions
-        const image = req.body.image
-        data = {
+        const image = req.body.image || "/images/default.jpg"
+
+        // declare data properly
+        const data = {
             title:title,
             ingredients:ingredients,
             instructions:instructions,
             image:image
         }
+
         //update using mongoose funciton 
         await Recipe.updateById(id,data)
+
         //redirect back to recipe in order to check if the change has been made
-        res.redirect("/recipe")
+        return res.redirect("/recipe")
     }
     catch(error){
         //placeholder before i res.render("error") if we still doing that 
-        res.send("You've got an error updating this recipe")
-
+        return res.send("You've got an error updating this recipe")
     }
 }
 
@@ -82,15 +132,29 @@ exports.updateRecipes = async (req,res) => {
 exports.deleteRecipes = async (req,res) =>{
     try{
         const id = req.params.id
-        const recipe = Recipe.findId(id)
-        if (!recipe){
-            res.send("No ID has been found")
+
+        //  must use await + correct function name
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.send("Invalid ID")
         }
+
+        const recipe = await Recipe.findById(id)
+
+        if (!recipe){
+            return res.send("No ID has been found")
+        }
+
         //authenticaiton placeholder
+        if (recipe.createdBy !== req.user.email){
+            return res.send("Not allowed")
+        }
 
         await Recipe.deleteById(id);
-        res.redirect("/recipe");
+
+        // redirects right back to recipe to see changes
+        return res.redirect("/recipe");
+
     }catch(error){
-        res.send("You got an error deleting this recipe")
+        return res.send("You got an error deleting this recipe")
     }
 }
